+++ lib/input/s3.go
@@ -21,8 +21,6 @@
 package input
 
 import (
-	"errors"
-
 	"github.com/Jeffail/benthos/lib/input/reader"
 	"github.com/Jeffail/benthos/lib/log"
 	"github.com/Jeffail/benthos/lib/metrics"
@@ -49,6 +47,11 @@ need to set the ` + "`sqs_body_path`" + ` field to where the object key is found
 in the payload. However, it is also common practice to send bucket events to an
 SNS topic which sends enveloped events to SQS, in which case you must also set
 the ` + "`sqs_envelope_path`" + ` field to where the payload can be found.
+
+When using SQS events it's also possible to extract target bucket names from the
+events by specifying a path in the field ` + "`sqs_bucket_path`" + `. For each
+SQS event, if that path exists and contains a string it will used as the bucket
+of the download instead of the ` + "`bucket`" + ` field.
 
 Here is a guide for setting up an SQS queue that receives events for new S3
 bucket objects:
@@ -89,9 +92,6 @@ You can access these metadata fields using
 
 // NewAmazonS3 creates a new AWS S3 input type.
 func NewAmazonS3(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
-	if len(conf.S3.Bucket) == 0 {
-		return nil, errors.New("invalid bucket (cannot be empty)")
-	}
 	r, err := reader.NewAmazonS3(conf.S3, log, stats)
 	if err != nil {
 		return nil, err